Renamed dbHelper to serverHelper
