+++ compare_courses/client/serverHelper.js
